+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to Q1Simulator will be documented in this file.
 
+## \[0.3.0] - 2022-04-11
+
+### Changed
+- Update for API changes of qblox_instruments v0.6
+   - Note: still backwards compatible with qblox_instruments v0.5
+- Instrument type 'QCM' or 'QRM' must be specified
+
 ## \[0.2.0] - 2022-03-07
 
 ### Fixed